requirements.txt
streamlit 
pdfplumber 
pandas 
PyPDF2 
msoffcrypto-tool 
openpyxl